+++ 3_Solution/mainwindow.h
@@ -1,0 +1,33 @@
+#ifndef MAINWINDOW_H
+#define MAINWINDOW_H
+
+#include <QMainWindow>
+#include <QtSql>
+#include <QSql>
+#include <QSqlQuery>
+#include <QSqlDatabase>
+#include <qmessagebox.h>
+
+
+QT_BEGIN_NAMESPACE
+namespace Ui { class MainWindow; }
+QT_END_NAMESPACE
+
+class MainWindow : public QMainWindow
+{
+    Q_OBJECT
+
+public:
+    MainWindow(QWidget *parent = nullptr);
+    ~MainWindow();
+
+private slots:
+    void on_LoginBtn_clicked();
+
+    void on_RegisterBtn_clicked();
+
+private:
+    Ui::MainWindow *ui;
+    QSqlDatabase database;
+};
+#endif // MAINWINDOW_H
