--- ATM_MAIL.pro
@@ -1,33 +1,0 @@
-QT       += core gui sql
-
-greaterThan(QT_MAJOR_VERSION, 4): QT += widgets
-
-CONFIG += c++17
-
-# You can make your code fail to compile if it uses deprecated APIs.
-# In order to do so, uncomment the following line.
-#DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
-
-SOURCES += \
-    mail_interface.cpp \
-    main.cpp \
-    mainwindow.cpp \
-    registerwindow.cpp
-
-HEADERS += \
-    mail_interface.h \
-    mainwindow.h \
-    registerwindow.h
-
-FORMS += \
-    mail_interface.ui \
-    mainwindow.ui \
-    registerwindow.ui
-
-# Default rules for deployment.
-qnx: target.path = /tmp/$${TARGET}/bin
-else: unix:!android: target.path = /opt/$${TARGET}/bin
-!isEmpty(target.path): INSTALLS += target
-
-RESOURCES += \
-    imagini.qrc
